requirements.txt
Adafruit-Blinka==8.66.2
Adafruit-Blinka-Raspberry-Pi5-Neopixel==1.0.0rc2
adafruit-circuitpython-busdevice==5.2.13
adafruit-circuitpython-connectionmanager==3.1.5
adafruit-circuitpython-dht==4.0.9
adafruit-circuitpython-requests==4.1.13
adafruit-circuitpython-typing==1.12.2
Adafruit-PlatformDetect==3.84.0
Adafruit-PureIO==1.1.11
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
attrs==25.4.0
binho-host-adapter==0.1.6
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorzero==2.0
decorator==5.2.1
frozenlist==1.8.0
future==1.0.0
geocoder==1.38.1
gpiozero==2.0.1
idna==3.10
lgpio==0.2.2.0
mido==1.3.3
multidict==6.7.0
packaging==25.0
pillow==11.3.0
propcache==0.4.0
pyftdi==0.57.1
pyserial==3.5
python-weather==2.1.0
pyusb==1.3.1
ratelim==0.1.6
requests==2.32.5
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
six==1.17.0
spidev==3.8
sysv-ipc==1.1.0
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.22.0
